✨ feat: add tabs
