--- views/dapp/v2/lsd/overview/overview.data.ts
@@ -1,14 +1,0 @@
-export const VALIDATORS_DATA = [
-  {
-    description: 'Validator',
-    length: 96,
-  },
-  {
-    description: 'Validator',
-    length: 96,
-  },
-  {
-    description: 'Validator',
-    length: 96,
-  },
-];
