+++ views/dapp/v2/lsd/staked-section/staking-form/index.tsx
