+++ views/dapp/v2/lsd/staked-section/staking-form/pro-mode/index.tsx
