+++ views/dapp/v2/lsd/staked-section/statistics/index.tsx
