+++ views/dapp/v2/lsd/staked-section/statistics/prices-market-card/trend-info.tsx
