+++ views/dapp/v2/lsd/staked-section/statistics/stats/index.tsx
