+++ views/dapp/v2/lsd/stats-section/overview/overview.data.ts
@@ -1,0 +1,19 @@
+import { SUISVG, UsersSVG } from '@/components/svg/v2';
+
+export const OVERVIEW_DATA = [
+  {
+    description: 'totalSuiStaked',
+    Icon: SUISVG,
+    value: 574.34,
+  },
+  {
+    description: 'totalStaked',
+    Icon: UsersSVG,
+    value: 96,
+  },
+  {
+    description: 'validators',
+    Icon: UsersSVG,
+    value: 96,
+  },
+];
